+++ lib/src/generator/templates.runtime_renderers.dart
@@ -25682,7 +25682,6 @@ const _invisibleGetters = {
     'flutterRoot',
     'hashCode',
     'include',
-    'includeExternal',
     'includeSource',
     'injectHtml',
     'inputDir',