+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 ## 9.0.0-wip
-* Remove deprecated `missingCodeBlockLanguage` warning.
+* Remove the deprecated `missingCodeBlockLanguage` warning.
 * Remove the deprecated `templates-dir` option.
 * Remove the deprecated `nodoc` option.
+* Remove the deprecated `include-external` option.
 
 ## 8.3.4
 * The URL for category pages now uses _category name_ instead of